upath._flavour: move _deprecated to upath._compat
